tweak tidy picking code
git-svn-id: http://vtp.googlecode.com/svn/trunk@7770 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/vtlib/core/Vehicles.cpp
@@ -201,7 +201,7 @@ void VehicleSet::AddEngine(CarEngine *e)
 	m_Engines.push_back(e);
 }
 
-int VehicleSet::FindClosestVehicle(const FPoint3 &point, float &closest)
+int VehicleSet::FindClosestVehicle(const FPoint3 &point, double &closest)
 {
 	closest = 1E9;
 	int vehicle = -1;
